Added effects and some bug fixes
Added a cpp file and a header file for effects called Effects.cpp/h. The files also rely on ysglfontdata.c/.h, which aren't included in the update. In order to get the program to run i added a default constructor to graduateStudent.cpp/.h as well as changed generateRand to generateRand1. I also implemented a getter for research in graduatestudent.cpp.